Extract duplicated methods to AbstractDependencyMojo
Moved resolveTargetProject(), existsInResolvedModel(), and
getModuleNames() from AddDependencyMojo and RemoveDependencyMojo
into the shared AbstractDependencyMojo base class. This eliminates
~40 lines of identical duplicated code and ensures consistent
behavior (e.g., RemoveDependencyMojo now includes available modules
in its error message, matching AddDependencyMojo).

Also cleaned up unused imports in both mojos.
